added the FOV on the minimap, some minor tweaks and a big refactor into new files and folders for clarity
