+++ .vscode/settings.json
@@ -2,6 +2,7 @@
 	"files.associations": {
 		"raycast.h": "c",
 		"init.h": "c",
-		"math.h": "c"
+		"math.h": "c",
+		"hooks.h": "c"
 	}
 }
