+++ incs/hooks.h
@@ -1,0 +1,19 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   hooks.h                                            :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: nponchon <nponchon@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/01/29 14:05:41 by nponchon          #+#    #+#             */
+/*   Updated: 2025/01/29 14:12:08 by nponchon         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef HOOKS_H
+# define HOOKS_H
+
+# include <math.h>
+# include "cub3d.h"
+
+#endif
