+++ incs/rendering.h
@@ -1,17 +1,17 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   raycast.h                                          :+:      :+:    :+:   */
+/*   rendering.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nponchon <nponchon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/25 12:39:04 by nponchon          #+#    #+#             */
-/*   Updated: 2025/01/28 14:45:11 by nponchon         ###   ########.fr       */
+/*   Updated: 2025/01/29 12:22:49 by nponchon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef RAYCAST_H
-# define RAYCAST_H
+#ifndef RENDERING_H
+# define RENDERING_H
 
 # include <math.h>
 # include "cub3d.h"
@@ -19,6 +19,6 @@
 void	cub_loop(void *param);
 int		cub_init_pixels(t_cub *cub);
 void	cub_draw(t_cub *cub);
-void	cub_draw_floor(t_cub *cub);
+void	cub_draw_background(t_cub *cub);
 
 #endif
