+++ src/main/cub3d.c
@@ -6,11 +6,44 @@
 /*   By: nponchon <nponchon@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/24 18:24:55 by nponchon          #+#    #+#             */
-/*   Updated: 2025/01/27 11:32:15 by nponchon         ###   ########.fr       */
+/*   Updated: 2025/01/29 14:07:26 by nponchon         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../incs/cub3d.h"
+
+/*
+	The main loop used in during the execution of cub3d and repeated
+	everytime the image needs to be refreshed. Updates the players position,
+	then renders the image.
+*/
+void	cub_loop(void *param)
+{
+	t_cub		*cub;
+
+	cub = param;
+	cub_hook(cub);
+	cub_draw(cub);
+}
+
+/*
+	Initialises the window and the image before starting the mlx loop.
+	Returns 1 if an error occured, 0 otherwise.
+*/
+int	cub_start(t_cub *cub)
+{
+	cub->mlx = mlx_init(WINDOW_WIDTH, WINDOW_HEIGHT, "Cub3D", true);
+	cub->img = mlx_new_image(cub->mlx, WINDOW_WIDTH, WINDOW_HEIGHT);
+	if (!cub->mlx || !cub->img)
+		return (1);
+	cub_draw(cub);
+	mlx_image_to_window(cub->mlx, cub->img, 0, 0);
+	mlx_loop_hook(cub->mlx, cub_loop, cub);
+	mlx_loop(cub->mlx);
+	mlx_delete_image(cub->mlx, cub->img);
+	mlx_terminate(cub->mlx);
+	return (0);
+}
 
 int	cub_cub3d(char **av)
 {
@@ -23,7 +56,7 @@ int	cub_cub3d(char **av)
 		return (1);
 	if (cub_parse(cub))
 		return (1);
-	if (cub_raycast(cub))
+	if (cub_start(cub))
 		return (1);
 	cub_clean(cub);
 	return (0);
